+++ R/read_airport_movements.R
@@ -10,6 +10,7 @@
 #'
 #' @template date
 #' @template showProgress
+#' @template cache
 #'
 #' @return A `"data.table" "data.frame"` object. All columns are returned with
 #'         `class` of type `"character"`.
@@ -21,7 +22,9 @@
 #'
 #' amov2020 <- read_airport_movements(date = 2020)
 #'}}
-read_airport_movements <- function(date = 202001, showProgress = TRUE){
+read_airport_movements <- function(date = 202001,
+                                   showProgress = TRUE ,
+                                   cache = TRUE){
 
   ### check inputs
   if( ! is.logical(showProgress) ){ stop(paste0("Argument 'showProgress' must be either 'TRUE' or 'FALSE.")) }
@@ -48,7 +51,9 @@ read_airport_movements <- function(date = 202001, showProgress = TRUE){
     file_url <- get_airport_movements_url(year=y, month=m)
 
     # download and read data
-    dt <- download_airport_movement_data(file_url, showProgress = showProgress)
+    dt <- download_airport_movement_data(file_url,
+                                         showProgress,
+                                         cache)
 
     # check if download failed
     if (is.null(dt)) { return(invisible(NULL)) } # nocov
@@ -85,7 +90,9 @@ read_airport_movements <- function(date = 202001, showProgress = TRUE){
                                     file_url <- get_airport_movements_url(year=y, month=m)
 
                                     # download and read data
-                                    temp_dt <- download_airport_movement_data(file_url, showProgress = FALSE)
+                                    temp_dt <- download_airport_movement_data(file_url,
+                                                                              showProgress = FALSE,
+                                                                              cache)
 
                                     # check if download failed
                                     if (is.null(temp_dt)) { return(invisible(NULL)) }